+++ bit/bit_test.go
@@ -51,16 +51,16 @@ func TestWriteReadRandomBitsLarge(t *testing.T) { testRandomBits(t, 1251) }
 func testRandomBits(t *testing.T, N int) {
 	values := make([]pair, N)
 	for i := range values {
-		width := uint(rand.Intn(61) + 1)
-		value := uint(rand.Intn(1<<width - 1))
+		width := uint64(rand.Intn(61) + 1)
+		value := uint64(rand.Intn(1<<width - 1))
 		values[i] = pair{value, width}
 	}
 	testReadWrite(t, values)
 }
 
 type pair struct {
-	value uint
-	width uint
+	value uint64
+	width uint64
 }
 
 func testReadWrite(t *testing.T, values []pair) {
