+++ bit/reverse_test.go
@@ -7,8 +7,8 @@ import (
 
 func TestReverseBitsRandom(t *testing.T) {
 	for i := 0; i < 1521; i += 1 {
-		width := uint(rand.Intn(61) + 1)
-		x := uint(rand.Intn(1<<width - 1))
+		width := uint64(rand.Intn(61) + 1)
+		x := uint64(rand.Intn(1<<width - 1))
 		rx := ReverseBits(x, width)
 		rrx := ReverseBits(rx, width)
 		if rrx != x {
@@ -18,8 +18,8 @@ func TestReverseBitsRandom(t *testing.T) {
 }
 
 func TestReverseBitsPowers(t *testing.T) {
-	for width := uint(2); width < 8; width += 1 {
-		exp := uint(1 << (width - 1))
+	for width := uint64(2); width < 8; width += 1 {
+		exp := uint64(1 << (width - 1))
 		got := ReverseBits(1, width)
 		if exp != got {
 			t.Errorf("fail <1,%v>: got %v exp %v", width, got, exp)
